Extract the index-swap machinery from `libqlever` into `index/IndexSwap.{h,cpp}`
Replacing the index at a given base name by a new index that was first
staged in a temporary subdirectory is a self-contained piece of on-disk
bookkeeping that has nothing to do with `libqlever`. Move it out of
`Qlever.{h,cpp}` into the new `index/IndexSwap.{h,cpp}`, so that it can
also be used from code that is not part of `libqlever`.

The moved entities, in detail:

* `qlever::IndexRebuildConfig` -> `qlever::IndexSwapConfig`, together
  with the file-local helper `baseNamesCollide`.
* The body of `Qlever::makeIndexRebuildConfig` -> the free function
  `qlever::makeIndexSwapConfig`, parameterized by the new struct
  `qlever::IndexSwapNaming`, which holds the prefixes for the two
  default directory names and the hint that is appended to the error
  message when they are all taken.
* The two lambdas `moveByBasename` and `moveIndex` of
  `Qlever::moveRebuiltIndexIntoPlace`, plus the removal of the staging
  directory -> the free function `qlever::moveIndexIntoPlace`.
* The member function `IndexRebuildConfig::successResponseAsJson()` ->
  the free function `qlever::rebuildSuccessResponseAsJson`, which stays
  in `libqlever` because it is about the HTTP response to a rebuild
  request, not about the swap.

`Qlever::makeIndexRebuildConfig` and `Qlever::moveRebuiltIndexIntoPlace`
remain as thin wrappers that supply the rebuild-specific naming and the
re-anchoring of the in-memory state of the new index.

Note on layering: `src/index` must not depend on `src/engine`, so
`moveIndexIntoPlace` enumerates the files of the materialized views with
`qlever::util::filesWithBaseNameAndSuffix(source, VIEW_FILE_INFIX)`
instead of `MaterializedViewsManager::viewFilesOnDisk`. This is what the
preceding commit introduced `VIEW_FILE_INFIX` in
`global/MaterializedViewConstants.h` for; the two are equivalent, as
`viewFilesOnDisk` is exactly that call.

This is a pure refactoring; no behavior changes. The one observable
reordering is that the (best-effort) removal of the now empty staging
directory happens before, instead of after, the in-memory state of the
new index is re-anchored. Both are independent of each other, and the
directory removal only logs on failure.
